pyproject.toml
[project]
name = "uc-intg-jvc"
version = "1.0.0"
description = "Unfolded Circle JVC Projector Integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.9.3",
    "defusedxml==0.7.1",
    "pydantic>=2.11.7",
    "pyee>=9.0",
    "pyjvcprojector==1.1.2",
    "requests>=2.32.4",
    "sddp-discovery-protocol==1.5.0",
    "ssdp==1.3.0",
    "ucapi==0.4.0",
    "ucapi-framework==1.0.1",
]
